Start big code refactoring. Moved gl initialization, gui init, camera init. Moved camera update.

+++ OpenGL/Camera.h
@@ -12,7 +12,7 @@ public:
 	~Camera();
 
 	void updateView();
-	void processInput(GLFWwindow* window, float deltaTime);
+	void processInput(GLFWwindow* window, double deltaTime);
 
 	inline const glm::mat4& viewMatrix() const { return m_viewMat; }
